refactoring utils, fix build
