+++ app/analytics/google-analytics.tsx
@@ -8,7 +8,7 @@ import React from 'react'
 
 import type { ScriptProps } from 'next/script'
 import Script from 'next/script' // not allow lazy load component
-import { logEnv } from './env'
+import { logEnv } from '@/utils/env'
 
 interface GoogleAnalyticsProps {
   gaMeasurementId?: string
