+++ app/utils/addHttpIfNeeded.ts
@@ -1,0 +1,6 @@
+export function addHttpIfNeed(url: string): string {
+  if (url.startsWith('http'))
+    return url
+
+  return `https://${url}`
+}
