+++ app/utils/env.ts
@@ -8,15 +8,8 @@ export function requireEnv(name: string, value: string | undefined): string {
   return value
 }
 
-export function addHttpIfNeed(url: string): string {
-  if (url.startsWith('http'))
-    return url
-
-  return `https://${url}`
-}
-
 /** Important to use process.env.[name] directly, because it replaced during build time */
-export function logEnv<T = undefined | string>(name: string, value: T): T {
+export function logEnv(name: string, value: undefined | string): undefined | string {
   // eslint-disable-next-line no-console
   console.log(name, value)
