+++ package.json
@@ -82,7 +82,7 @@
     "redux": "^4.2.1",
     "sharp": "^0.33.5",
     "siwe": "^2.3.2",
-    "typescript": "^5.2.2",
+    "typescript": "5.3.3",
     "uuid": "^10.0.0"
   },
   "devDependencies": {
